requirements.txt
absl-py==2.1.0
addict==2.4.0
aiodns==3.1.1
aiofiles==23.2.1
aiohttp==3.9.4
aioitertools==0.11.0
aiosignal==1.3.1
alphashape==1.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
asteval==0.9.32
asttokens==3.0.0
async-timeout==4.0.3
attrs==23.2.0
azure-core==1.28.0
azure-identity==1.13.0
azure-storage-blob==12.17.0
backcall==0.2.0
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.9.0
casadi==3.7.0
cchardet==2.1.7
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
click-log==0.4.0
cloudpickle==3.1.1
comm==0.2.2
ConfigArgParse==1.7
construct==2.10.68
contourpy==1.3.1
coverage==7.4.4
cryptography==42.0.7
cycler==0.11.0
dash==2.18.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==4.4.2
defusedxml==0.7.1
diffusers==0.33.1
docopt==0.6.2
einops==0.8.1
evdev==1.9.2
exceptiongroup==1.2.2
executing==2.2.0
Farama-Notifications==0.0.4
fastapi==0.110.0
fastjsonschema==2.21.1
filelock==3.13.1
Flask==3.0.3
fonttools==4.56.0
frozenlist==1.4.1
fsspec==2024.2.0
gitdb==4.0.12
GitPython==3.1.44
grpcio==1.70.0
gunicorn==22.0.0
gymnasium==0.29.0
h11==0.14.0
h5py==3.13.0
hidapi==0.14.0.post2
httptools==0.6.1
huggingface-hub==0.29.2
hydra-core==1.3.2
idna==3.7
idna-ssl==1.1.0
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.6.1
ipython==8.12.3
ipywidgets==8.1.5
isaaclab==2.0.2
isodate==0.6.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
jupyterlab_widgets==3.0.13
kiwisolver==1.4.4
lazy_loader==0.4
llvmlite==0.42.0
lxml==5.2.2
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mathutils==3.3.0
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
moviepy==1.0.3
mpmath==1.3.0
msal==1.23.0
msal-extensions==1.0.0
multidict==6.0.5
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.5.6
networkx==3.2.1
ninja==1.11.1.3
numba==0.59.1
numpy==1.26.0
numpy-quaternion==2023.0.3
oauthlib==3.2.2
omegaconf==2.3.0
onnx==1.16.1
open3d==0.19.0
opencv-python==4.11.0.86
osqp==0.6.5
packaging==23.0
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pathfinding==1.0.16
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.0.0
Pint==0.20.1
pipreqs==0.5.0
platformdirs==4.3.6
plotly==5.3.1
portalocker==2.7.0
prettytable==3.3.0
proglog==0.1.10
prompt_toolkit==3.0.50
protobuf==4.25.6
psutil==5.9.8
ptyprocess==0.7.0
pure_eval==0.2.3
pycares==4.3.0
pycocotools==2.0.8
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pyglet==1.5.31
Pygments==2.19.1
pynput==1.8.1
pyparsing==3.0.9
pyperclip==1.8.0
pypng==0.20220715.0
pyquaternion==0.9.9
python-dateutil==2.9.0.post0
python-multipart==0.0.9
python-xlib==0.33
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.4.0
qdldl==0.1.7.post1
qrcode==7.4.2
referencing==0.36.2
regex==2024.11.6
requests==2.31.0
requests-oauthlib==1.3.1
retrying==1.3.4
rich==14.0.0
rpds-py==0.22.3
rsl-rl-lib==2.3.1
rtree==1.4.0
s3transfer==0.6.1
safetensors==0.5.3
scikit-image==0.25.1
scikit-learn==1.6.1
scipy==1.11.2
selenium==4.14.0
sentry-sdk==1.43.0
setuptools-scm==8.0.4
shapely==2.0.7
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
stack-data==0.6.3
starlette==0.36.3
sympy==1.13.1
tenacity==9.0.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
termcolor==3.0.1
terminaltables==3.1.10
threadpoolctl==3.5.0
tifffile==2025.1.10
tinycss2==1.4.0
tokenizers==0.21.0
toml==0.10.2
tomli==2.0.1
tornado==6.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.49.0
trimesh==4.6.4
triton==3.1.0
typing_extensions==4.11.0
tzdata==2025.1
urllib3==2.2.2
usd-core==25.5.1
uvicorn==0.29.0
warp-lang==1.6.1
watchdog==4.0.0
wcwidth==0.2.13
webbot==0.34
webencodings==0.5.1
websockets==12.0
Werkzeug==3.0.6
widgetsnbextension==4.0.13
wrapt==1.16.0
yapf==0.43.0
yarg==0.1.9
yarl==1.9.4
zipp==3.21.0
